Anim BP -> C++
애니메이션 인스턴스 C++로 대체

+++ Source/DreamWorld/Animation/Player/AI_Player.h
@@ -31,5 +31,5 @@ private:
 	FVector							m_Velocity;
 
 	UPROPERTY(Transient, EditAnywhere, BlueprintReadOnly, Category = "Pawn | Speed", meta = (AllowPrivateAccess = "true"))
-	float									m_Speed;
+	double								m_Speed;
 };
